Major refactor
An attempt to simplify the directory structure and file naming, it was
all getting a bit complex with local requests, supplier requests and
actions.

Also, I have removed any references to actions. ISO18626 is pretty clear
that we have "Request" and then subsequent "Messages", we should use
this terminology going forward. But, for now, we are only catering for
"Request"