Remove request confirmation ErrorData enum
I'm not at all sure about this as it diverges from the ISO spec. This
commit removes the enum of possible error messages that can be supplied
in ErrorData. As far as I can see, they are pretty useless. There is no
way to convey actual error messages, for example, from a supplier, such
as "System unavailable" or suchlike. It is now therefore possible to
send any arbitrary string in ErrorData :-/

I _did_ consider keeping the enum and adding a header to convey any
error messages, but even then it's not possible to select a correct enum
value most of the time since they're so specific. We need an
"UndefinedError" value in the enum!