Remove date-time type from date fields
Following this discussion:

https://folio-project.slack.com/archives/CC0PHKEMT/p1635936905024600

It seems that there is currently no way to automatically serialize and
deserialize java.time type objects as they pass in and out of the
module.

This will be possible at some point (I'm hoping to follow Julian's
suggestion and create an RMB PR), but for now manual serialisation and
deserialisation will be necessary.